Converting variables types to size_t from int. Causing github check w… (#4902)
* Converting variables types to size_t from int. Causing github check warnings. Technically cast from size_t to int can have an overflow/loss of data but very unlikely in these scenarios.

* Changing variable types to size_t from int. Can have overflow issues.

* fix some more msvc warnings

* fix some more warnings from msvc

+++ protocol/server/handle_client_message.c
@@ -354,7 +354,7 @@ static int handle_file_metadata_message(WhistClientMessage *wcmsg) {
 
     static int local_directory_id = 0;  // static so only set to 0 on first iteration
     static const char *parent_drag_drop_directory = "/home/whist/.teleport/drag-drop";
-    int target_directory_maxlen =
+    size_t target_directory_maxlen =
         strlen(parent_drag_drop_directory) + 64 + 1;  // have enough space for int in paths
 
     // Use `unique_id` for the rest of the function, NOT `local_directory_id`.
